requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
layoutparser==0.3.4
pytesseract==0.3.13
Pillow==11.3.0
numpy==2.2.6
pandas==2.3.2
python-dotenv==1.1.1
pdf2image==1.17.0
scikit-image==0.24.0
scipy==1.16.1
